% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{State map of the year data fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Number of a state}

\item{year}{The year}
}
\value{
A plot from the data file
}
\description{
Accepts a state number and year
Makes the filename using make_filename function
gets a data frame from fars_read()
}
\details{
Error checks to make sure the state number exists
If so, uses maps and graphics to create plots based on latitude and longitude from the data file

Uses make_filename(year)
     fars_read(filename)
}
\examples{
\dontrun{
fars_map_state(1, '2013')
}
}
