% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read years fars_read_years}
\source{
fars_data/data/accident_year.csv.bz2
}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{the year we need to get the data file}
}
\value{
Creates datasets based on year.  Returns NULL if there is an error
}
\description{
This function accepts year and returns data of the associated year.
}
\examples{
\dontrun{
    fars_read_years(c(2013, 2014))
    fars_read_years(c("2013", "2014"))
          }
}
