% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Years fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{the year we need to get the data file}
}
\value{
summary of the data  associated with the year
}
\description{
This function accepts year and summarize data of the associated year.
}
\examples{
  \dontrun{
    fars_summarize_years(c(2013, 2014))
    fars_summarize_years(c("2013", "2014"))
  }
}
