% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year we need to get the data file}
}
\value{
name of the data file associated with the year
}
\description{
This function accepts year and returns the file name of the associated year.
}
\examples{
\dontrun{
x<- make_filename('1979')
}
}
