\name{newletter}
\alias{newletter}
test
\title{
test
}
\description{
test
}
\usage{
newletter(x)
}
test
\arguments{
  \item{x}{
test
}
}
\details{
test
}
\value{
test
}
\references{
test
}
\author{
test
}
\note{
test
}

test

\seealso{
test
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x) 
{
    rando <- sample(letters[1:26], 1)
    return(rando)
  }
}
test
\keyword{ ~kwd1 }test
\keyword{ ~kwd2 }test
